# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "GinBlog博客系统API接口文档",
        "title": "GinBlog博客系统",
        "contact": {},
        "version": "1.0.1"
    },
    "paths": {
...
        "/api/GetProvinceAccess": {
            "get": {
                "description": "返回各省份的访问总量、最后访问时间，支持分页和排序",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "搜索统计相关接口"
                ],
                "summary": "查询各省份访问量统计",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码（默认1，最小1）",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数（默认20，最大100）",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序字段（access_count/last_access_at，默认access_count）",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序方式（asc/desc，默认desc）",
                        "name": "sort_order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/result.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/model.ProvincePageResponse"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "list": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/model.ProvinceAccessStatisticResp"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/result.Result"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/result.Result"
                        }
                    }
                }
            }
        },
...
        "/api/GetTagSearch": {
            "get": {
                "description": "关联标签表与标签搜索统计表，返回各标签的名称、搜索次数、最后搜索时间（未被搜索过的标签搜索量为0）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "搜索统计相关接口"
                ],
                "summary": "查询标签搜索量统计",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码（默认1，最小1）",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数（默认20，最大100）",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序字段（search_count/last_search_at，默认search_count）",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序方式（asc/desc，默认desc）",
                        "name": "sort_order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功\" // 这里必须加model.前缀",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/result.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/model.TagPageResponse"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "list": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/model.TagSearchStatisticResp"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/result.Result"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/result.Result"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.ProvinceAccessStatisticResp": {
            "type": "object",
            "properties": {
                "access_count": {
                    "description": "该省份总访问量",
                    "type": "integer"
                },
                "last_access_at": {
                    "description": "该省份最后访问时间",
                    "allOf": [
                        {
                            "$ref": "#/definitions/utils.HTime"
                        }
                    ]
                },
                "province": {
                    "description": "省份名称（如：广东）",
                    "type": "string"
                }
            }
        },
        "model.ProvincePageResponse": {
            "type": "object",
            "properties": {
                "list": {
                    "description": "地区统计列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ProvinceAccessStatisticResp"
                    }
                },
                "page": {
                    "description": "当前页码",
                    "type": "integer"
                },
                "page_size": {
                    "description": "每页条数",
                    "type": "integer"
                },
                "total": {
                    "description": "总省份数",
                    "type": "integer"
                },
                "total_page": {
                    "description": "总页数",
                    "type": "integer"
                }
            }
        },
...
        "model.TagPageResponse": {
            "type": "object",
            "properties": {
                "list": {
                    "description": "用 interface{} 支持任意列表类型"
                },
                "page": {
                    "description": "当前页码",
                    "type": "integer"
                },
                "page_size": {
                    "description": "每页条数",
                    "type": "integer"
                },
                "total": {
                    "description": "总条数（数据库count返回int64）",
                    "type": "integer"
                },
                "total_page": {
                    "description": "总页数（int类型）",
                    "type": "integer"
                }
            }
        },
        "model.TagSearchStatisticResp": {
            "type": "object",
            "properties": {
                "last_search_at": {
                    "description": "最后搜索时间（未搜索过为null）",
                    "allOf": [
                        {
                            "$ref": "#/definitions/utils.HTime"
                        }
                    ]
                },
                "search_count": {
                    "description": "搜索次数（未搜索过为0）",
                    "type": "integer"
                },
                "tag_name": {
                    "description": "标签名称",
                    "type": "string"
                }
            }
        },
...
        "result.Result": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
        "utils.HTime": {
            "type": "object",
            "properties": {
                "time.Time": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}